Move ChangeKioskController to admin package
