+++ src/controllers/admin/AdminMenuController.java
@@ -1,19 +1,11 @@
 package controllers.admin;
 
 import controllers.AbstractController;
-import controllers.ChangeKioskController;
-import controllers.DirectoryViewController;
 import controllers.MainMenuController;
 import core.KioskMain;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
 import javafx.scene.control.Button;
-import javafx.stage.Stage;
-
-import java.io.IOException;
 
 /**
  * Created by mattm on 3/29/2017.
